fix package policy being able save support_cloud_connector error
